fix cycling refer
